backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
Click==7.0
coverage==5.0.1
Django==3.0.1
django-cors-headers==3.2.0
djangorestframework==3.11.0
pathspec==0.6.0
pytz==2019.3
regex==2019.12.20
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
